+++ lib/data/repository/user_repository.dart
@@ -6,20 +6,20 @@ import 'package:project_coffee/data/mock/user.dart';
 
 class UserRepository {
   Future<ResponseDTO> fetchJoin(JoinReqDTO joinReqDTO) async {
-    try {
-      // dynamic -> http body
-      Response<dynamic> response =
-          await dio.post("/join", data: joinReqDTO.toJson());
-      ResponseDTO responseDTO = ResponseDTO.fromJson(response.data);
-
-      //responseDTO.data = User.fromJson(responseDTO.data);
-
-
-      return responseDTO;
-    } catch (e) {
-      // 200이 아니면 catch로 감
-      return ResponseDTO(false, "중복되는 유저명입니다", null);
-    }
+    // try {
+    //   // dynamic -> http body
+    //   Response<dynamic> response =
+    //       await dio.post("/join", data: joinReqDTO.toJson());
+    //   ResponseDTO responseDTO = ResponseDTO.fromJson(response.data);
+    //
+    //   //responseDTO.data = User.fromJson(responseDTO.data);
+    //
+    //
+    //   return responseDTO;
+    // } catch (e) {
+    //   // 200이 아니면 catch로 감
+    //   return ResponseDTO(false, "중복되는 유저명입니다", null);
+    // }
     return Future.delayed(Duration(seconds: 3), () => mUserResponseDTO);
   }
