+++ lib/ui/pages/home_page/login_page/widget/login_page_body.dart
@@ -12,8 +12,8 @@ class LoginPageBody extends StatelessWidget {
     return CustomScrollView(
       slivers: [
         CustomSliverAppBar(title: "로그인",),
-
-        form,
+        // form,
+        LoginPageBodyItem(),
 
       ],
     );
