+++ lib/ui/pages/home_page/login_page/widget/login_page_body_item.dart
@@ -11,6 +11,9 @@ import 'package:project_coffee/data/store/session_store.dart';
 import 'package:project_coffee/ui/pages/main_page/main_page.dart';
 import 'package:project_coffee/ui/widgets/custom_text_form_field.dart';
 
+import 'login_page_find_and_join.dart';
+import 'login_page_form_field.dart';
+
 class LoginPageBodyItem extends ConsumerWidget {
   final _formKey = GlobalKey<FormState>();
   final userId = TextEditingController();
@@ -21,7 +24,6 @@ class LoginPageBodyItem extends ConsumerWidget {
   });
 
   void submit(WidgetRef ref) {
-    print("여기봐라~~ ${_formKey.currentState}");
     if (_formKey.currentState!.validate()) {
       LoginReqDTO loginReqDTO = LoginReqDTO(
         userId: userId.text,
@@ -53,20 +55,7 @@ class LoginPageBodyItem extends ConsumerWidget {
             SizedBox(height: gap_m,),
             Text("회원 서비스 이용을 위해 로그인 해주세요."),
             SizedBox(height: gap_xl + 5,),
-            Form(
-                key: _formKey,
-                child:
-            Column(
-              children: [
-                CustomTextForm("UserId", validatorFunction: validateUserId,
-                  controller: userId,),
-                SizedBox(height: gap_l,),
-                CustomTextForm("Password", validatorFunction: validatePassword,
-                  controller: password,),
-                SizedBox(height: gap_xl,),
-              ],
-            )
-            ),
+            LoginPageFormField(formKey: _formKey, userId: userId, password: password),
 
             SizedBox(height: gap_l,),
             FindAndJoin()
@@ -78,35 +67,6 @@ class LoginPageBodyItem extends ConsumerWidget {
 }
 
 
-class FindAndJoin extends StatelessWidget {
-  const FindAndJoin({
-    super.key,
-  });
 
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        TextButton(onPressed: () {
-          //컨벤션 회의해야함
-          // Navigator.push(context, MaterialPageRoute(builder: (context) => FindLoginIdPage()),);
-          Navigator.pushNamed(context, Move.FindLoginIdPage);
-        }, child: Text("아이디 찾기", style: TextStyle(color: Colors.black),)),
-        Container(
-          width: 1,
-          color: Colors.black26,
-        ),
-
-        TextButton(onPressed: () {
-          Navigator.pushNamed(context, Move.FindPasswordPage);
-        }, child: Text("비밀번호 찾기", style: TextStyle(color: Colors.black),)),
-        TextButton(onPressed: () {
-          Navigator.pushNamed(context, Move.JoinPage);
-        }, child: Text("회원가입", style: TextStyle(color: Colors.black),)),
-      ],
-    );
-  }
-}
