+++ lib/ui/widgets/custom_green_button.dart
@@ -36,8 +36,6 @@ class CustomGreenButton extends StatelessWidget {
                 MaterialPageRoute(builder: (context) => move),
               );
             }
-
-
           },
           child: Text(
             "${title}",
